Update README.md
This code is intended to analyze the sensitivity of functional ultrasound and functional MRI data acquired in anesthetized mice in response to different intensities of optogenetic stimulation of the primary motor cortex. Analyses involve preprocessing and general linear models to assess neural circuit activation. Additional scripts are included to analyze local field potential data acquired in the bilateral primary motor cortex and striatum.